+++ src/main/gen/io/vlang/lang/psi/impl/VlangMethodDeclarationImpl.java
@@ -65,9 +65,9 @@ public class VlangMethodDeclarationImpl extends VlangMethodDeclarationWithScopeH
   }
 
   @Override
-  @NotNull
+  @Nullable
   public VlangSignature getSignature() {
-    return notNullChild(VlangPsiTreeUtil.getStubChildOfType(this, VlangSignature.class));
+    return VlangPsiTreeUtil.getStubChildOfType(this, VlangSignature.class);
   }
 
   @Override