+++ src/test/resources/parser/new/attributes.v
@@ -11,11 +11,11 @@ struct Name {
 	name string @[sql]
 	age  int    @[default: 0]
 	tag  string @[default: nil]
-	baz  bool   @[prefix.attr0] = false
+	baz  bool = false   @[prefix.attr0]
 	name1 string @[dot.name]
 	name2 string @[name: 100]
 	name3 string @[sql: 100]
-	name4 string @[unsafe: 100]
+	name4 string @[unsafe; value: 100]
 }
 
 @[minify]